show all images from s3 on the main page for logged in users
