+++ aws_functions.py
@@ -3,6 +3,7 @@ Call AWS services from here!
 """
 import boto3
 import utils
+from image import Image
 from botocore.exceptions import ClientError
 
 BUCKET_NAME = 'photosharingapp-test'
@@ -44,8 +45,19 @@ def get_all_images(username):
     :param username: owner's name associated with images
     :return: array of strings, names of all images for this user
     """
-    # TODO: return list of image_id (array of strings) for this user
-    pass
+    # TODO: retrieve information from DynamoDB and create array of objects of type Image
+    # TODO: Remove everything below after implementing DynamoDB version
+    # Retrieving all files from S3 and faking other data for Image object
+    s3 = boto3.client('s3')
+    resp = s3.list_objects_v2(Bucket=BUCKET_NAME)
+    images_data = list()
+    for obj in resp['Contents']:
+        image_id = obj["Key"]
+        description = "Description for image " + image_id + " and some smart text here! #AWS #working"
+        images_data.append(Image(image_id=image_id, username=username, description=description, tags=list(), privacy=True))
+    print(len(images_data))
+    return images_data
+
 
 
 def store_image_data(image_binary, username, description, privacy):
@@ -62,7 +74,7 @@ def store_image_data(image_binary, username, description, privacy):
     image_id = utils.create_picID(username)
     # store image on AWS S3
     store_image_object(image_id, image_binary)
-    # TODO: Write DynamoDB code storing all the image information (do not use/store image_binary in DynamoDB)
+    # TODO: Write DynamoDB code storing all the image information (you don't need to use image_binary in DynamoDB, it is stored in S3)
     # split tags
     tags = utils.split_by_tag(description)
     # store image_id, description, tags, username, etc on DynamoDB
